backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
cffi==1.15.0
Django==4.0.3
djangorestframework==3.13.1
pycparser==2.21
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
